requirements.txt
geopy==2.2.0
numpy==1.23.3
requests==2.28.1
simpy==4.0.1
azure.storage.blob
azure.identity
azure.keyvault.secrets